requirements.txt
contourpy==1.2.1
cycler==0.12.1
ffmpeg-python==0.2.0
fonttools==4.51.0
future==1.0.0
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pip==24.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
setuptools==69.5.1
six==1.16.0
tqdm==4.66.2
wheel==0.43.0
